+++ CHANGELOG.md
@@ -24,6 +24,8 @@ The deprecated model `simple_aperture` in `photon_collection` is removed.
   and [#643](https://gitlab.com/esa/pyxel/-/issues/643)).
 * Fix bug in function `pyxel.run` when saving YAML configuration file.
   (See [!843](https://gitlab.com/esa/pyxel/-/merge_requests/843)).
+* Fix issue with `pyxel.run_mode()` in Calibration mode.
+  (See [!844](https://gitlab.com/esa/pyxel/-/merge_requests/844)).
 
 ### Documentation
 * Update documentation about quick start installation.